Pickup Component and BodyPartObject
Added current_object_held to the eventbus, this should be probably be moved into a different global. This is used in the Pickup Component to stop multiple objects from being picked up.

Created a pickup component and bodypart object which will be used in the build scene.

Started making an object slot to hook into the ui letting you place bodypart objects into the slot and then access them to create a new unit.

I put the object slot into an "unorganized" folder because I didn't really know where to put it