removed UnitLeader and started new group signal
UnitLeader didn't really make sense, so I just put a level property on UnitGroup. I am still trying to figure out how exactly units make up a unit group.

Started a signal to notify the HUD and the group manager when a new group is created with the debug menu